added confirmation when deleting a scene

+++ webapp/static/js/all_annoscenes.js
@@ -25,6 +25,13 @@ let all_annoscene_table = new Vue({
         delete_scene: function (scene_idx) {
             let self = this;
             let scene_id = self.scenes[scene_idx]._id;
+            let scene_name = self.scenes[scene_idx].name;
+
+            // popup a message box to let user confirm whether to delete
+            if (!confirm("Are you sure to delete scene " + scene_name + "?")) {
+                return;
+            }
+
             axios.delete('/annoscene/'+scene_id)
                 .then(function (response) {
                     console.log("Scene " + scene_id + " has been deleted.")
